+++ packages/bootstrap-4/src/ArrayFieldItemTemplate/ArrayFieldItemTemplate.tsx
@@ -16,6 +16,7 @@ const ArrayFieldItemTemplate = (props: ArrayFieldTemplateItemType) => {
     onReorderClick,
     readonly,
     registry,
+    uiSchema,
   } = props;
   const { MoveDownButton, MoveUpButton, RemoveButton } =
     registry.templates.ButtonTemplates;
@@ -41,6 +42,7 @@ const ArrayFieldItemTemplate = (props: ArrayFieldTemplateItemType) => {
                     style={btnStyle}
                     disabled={disabled || readonly || !hasMoveUp}
                     onClick={onReorderClick(index, index - 1)}
+                    uiSchema={uiSchema}
                   />
                 </div>
               )}
@@ -50,6 +52,7 @@ const ArrayFieldItemTemplate = (props: ArrayFieldTemplateItemType) => {
                     style={btnStyle}
                     disabled={disabled || readonly || !hasMoveDown}
                     onClick={onReorderClick(index, index + 1)}
+                    uiSchema={uiSchema}
                   />
                 </div>
               )}
@@ -59,6 +62,7 @@ const ArrayFieldItemTemplate = (props: ArrayFieldTemplateItemType) => {
                     style={btnStyle}
                     disabled={disabled || readonly}
                     onClick={onDropIndexClick(index)}
+                    uiSchema={uiSchema}
                   />
                 </div>
               )}