+++ packages/bootstrap-4/src/ArrayFieldTemplate/ArrayFieldTemplate.tsx
@@ -81,6 +81,7 @@ const ArrayFieldTemplate = (props: ArrayFieldTemplateProps) => {
                       className="array-item-add"
                       onClick={onAddClick}
                       disabled={disabled || readonly}
+                      uiSchema={uiSchema}
                     />
                   </Col>
                 </Row>