+++ packages/chakra-ui/src/AddButton/AddButton.tsx
@@ -3,7 +3,10 @@ import { IconButtonProps } from "@rjsf/utils";
 import { Button } from "@chakra-ui/react";
 import { AddIcon } from "@chakra-ui/icons";
 
-const AddButton: React.ComponentType<IconButtonProps> = (props) => (
+const AddButton: React.ComponentType<IconButtonProps> = ({
+  uiSchema,
+  ...props
+}) => (
   <Button leftIcon={<AddIcon />} {...props}>
     Add Item
   </Button>