+++ packages/chakra-ui/src/ArrayFieldTemplate/ArrayFieldTemplate.tsx
@@ -73,6 +73,7 @@ const ArrayFieldTemplate = (props: ArrayFieldTemplateProps) => {
                 className="array-item-add"
                 onClick={onAddClick}
                 disabled={disabled || readonly}
+                uiSchema={uiSchema}
               />
             </Box>
           </GridItem>