+++ packages/material-ui/src/ArrayFieldItemTemplate/ArrayFieldItemTemplate.tsx
@@ -16,6 +16,7 @@ const ArrayFieldItemTemplate = (props: ArrayFieldTemplateItemType) => {
     onDropIndexClick,
     onReorderClick,
     readonly,
+    uiSchema,
     registry,
   } = props;
   const { MoveDownButton, MoveUpButton, RemoveButton } =
@@ -44,6 +45,7 @@ const ArrayFieldItemTemplate = (props: ArrayFieldTemplateItemType) => {
               style={btnStyle}
               disabled={disabled || readonly || !hasMoveUp}
               onClick={onReorderClick(index, index - 1)}
+              uiSchema={uiSchema}
             />
           )}
           {(hasMoveUp || hasMoveDown) && (
@@ -51,6 +53,7 @@ const ArrayFieldItemTemplate = (props: ArrayFieldTemplateItemType) => {
               style={btnStyle}
               disabled={disabled || readonly || !hasMoveDown}
               onClick={onReorderClick(index, index + 1)}
+              uiSchema={uiSchema}
             />
           )}
           {hasRemove && (
@@ -58,6 +61,7 @@ const ArrayFieldItemTemplate = (props: ArrayFieldTemplateItemType) => {
               style={btnStyle}
               disabled={disabled || readonly}
               onClick={onDropIndexClick(index)}
+              uiSchema={uiSchema}
             />
           )}
         </Grid>